@@ -34,7 +34,7 @@ import { WidgetSettings, WidgetSettingsComponent } from '@shared/models/widget.m
 import { UntypedFormBuilder, UntypedFormGroup } from '@angular/forms';
 import { Store } from '@ngrx/store';
 import { AppState } from '@core/core.state';
-import { MatChipInputEvent, MatChipList } from '@angular/material/chips';
+import { MatChipInputEvent, MatChipGrid } from '@angular/material/chips';
 import { MatAutocomplete, MatAutocompleteSelectedEvent } from '@angular/material/autocomplete';
 import { COMMA, ENTER, SEMICOLON } from '@angular/cdk/keycodes';
 import { Observable, of, Subject } from 'rxjs';
@@ -47,7 +47,7 @@ import { map, mergeMap, share, startWith } from 'rxjs/operators';
 })
 export class NavigationCardsWidgetSettingsComponent extends WidgetSettingsComponent {
 
-  @ViewChild('filterItemsChipList') filterItemsChipList: MatChipList;
+  @ViewChild('filterItemsChipList') filterItemsChipList: MatChipGrid;
   @ViewChild('filterItemAutocomplete') filterItemAutocomplete: MatAutocomplete;
   @ViewChild('filterItemInput') filterItemInput: ElementRef<HTMLInputElement>;
 
