@@ -39,7 +39,7 @@ import { Edge } from "@shared/models/edge.models";
 @Component({
   selector: 'tb-edge-table-header',
   templateUrl: './edge-table-header.component.html',
-  styleUrls: ['./edge-table-header.component.scss']
+  styleUrls: []
 })
 export class EdgeTableHeaderComponent extends EntityTableHeaderComponent<Edge> {
 
