@@ -33,6 +33,7 @@ import { ChangeDetectorRef, Component, EventEmitter, Input, Output } from '@angu
 import { ClipboardService } from 'ngx-clipboard';
 import { TooltipPosition } from '@angular/material/tooltip';
 import { TranslateService } from '@ngx-translate/core';
+import { ThemePalette } from '@angular/material/core';
 
 @Component({
   selector: 'tb-copy-button',
@@ -67,7 +68,7 @@ export class CopyButtonComponent {
   style: {[key: string]: any} = {};
 
   @Input()
-  color: string;
+  color: ThemePalette;
 
   @Output()
   successCopied = new EventEmitter<string>();
